Further enhance list-like API to support adding properties. Update docs.
Former-commit-id: 62f284c813d794a8e68b4eb6a20c5a5892e33ec8 [formerly 888c026248699fd8b3a6c556b2dc120c7047320c]
Former-commit-id: 5a08549696cf6b2b8e57c6a85c8a2ac6ecf9d26f

+++ pymatgen/io/vaspio/vasp_output.py
@@ -1717,7 +1717,7 @@ class VolumetricData(object):
         inds = data[:, 1] <= radius
         dists = data[inds, 1]
         data_inds = np.rint(np.mod(list(data[inds, 0]), 1) *
-                            np.tile(a, (len(dists), 1)))
+                            np.tile(a, (len(dists), 1))).astype(int)
         vals = [self.data["diff"][x, y, z] for x, y, z in data_inds]
 
         hist, edges = np.histogram(dists, bins=nbins,
